[sync] split implementation into index and impl.js
